refactor: move notification setting out of setting page
