+++ lib/presentation/notification/providers/notification_setting_use_case_provider.dart
@@ -1,12 +1,10 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-
 import 'package:dongsoop/domain/notification/use_case/notification_disable_use_case.dart';
 import 'package:dongsoop/domain/notification/use_case/notification_enable_use_case.dart';
 import 'package:dongsoop/domain/notification/use_case/notification_setting_use_case.dart';
 import 'package:dongsoop/domain/notification/use_case/notification_setting_apply_use_case.dart';
 import 'package:dongsoop/domain/notification/use_case/notification_setting_result_use_case.dart';
-
-import 'package:dongsoop/presentation/setting/notification/providers/notification_setting_provider.dart';
+import 'package:dongsoop/presentation/notification/providers/notification_setting_provider.dart';
 
 final notificationSettingUseCaseProvider =
 Provider<NotificationSettingUseCase>((ref) {
